[rb] Fix "no anonymous block parameter" in ruby 3.1 (#15315)
Fix "no anonymous block parameter" in ruby 3.1

When running on Ruby 3.1.2p20 (shipped with Debian 12), the use of
anonymous blocks cause exceptions to be raised.  This was fixed in later
Ruby versions, but will not be backported to the version of Ruby shipped
by Debian 12 nor the versions or ruby shipped with all Debian
derivative.

Explicitly name blocks to avoid triggering this issue with affected
versions of Ruby.

While here, also adjust the rubocop configuration to detect this issue
and mandate the use of explicitly named blocks.

Signed-off-by: Romain Tartière <romain@blogreen.org>
Co-authored-by: Viet Nguyen Duc <nguyenducviet4496@gmail.com>