[grid] Silent fail on invalid log level (#15796)
* [grid] Silent fail on invalid log level

Fixing issue where providing an invalid log level to Grid will fail silently.

Since the logger is not yet configured a stacktrace with the exception is printed to console and we continue since the default log level is already configured.

Fixes #15790

* [grid] Silent fail on invalid log level

Adding link to logging levels as suggested by @diemol and formatting.

* [grid] Silent fail on invalid log level

Adding hardcoded log levels as suggested by review.