[dotnet] Enabling drivers to set log to console. (#16097)
* [dotnet] Enabling drivers to set log to console.

This will be false by default, following what the

.NET bindings have usually done.

Helps with #12273

* [dotnet] Avoid changing state of variable.

LogPath takes precedence anyway.

* Using existing logger functionality.

Messages will be shown to the console
using the Logger functionality.

Adding a test that listens to the
logger and checks that logs are being
sent to console.

Fixing an issue when sending the log
to a file, as the OnDriverProcessStarted
method in the base class was not being
invoked.

* Running format script.

* WriteDriverLogToConsole

Internal variable to avoid using
the stream when logs are sent to a file.

Not exposed to the user.

* Format script.

* Update dotnet/src/webdriver/DriverService.cs

Co-authored-by: Nikolay Borisenko <22616990+nvborisenko@users.noreply.github.com>

* Switching from async tasks to sync threads.

* Format script.

* Reading output through async events.

* Addressing PR comments

---------

Co-authored-by: Nikolay Borisenko <22616990+nvborisenko@users.noreply.github.com>